[Clean] 사용하지 않는 변수 및 스크립트 제거
-사용하지 않는 변수 및 스크립트 제거
-조그만한 리펙토링 및 의미없는 캐싱 삭제
-스크립트 폴더 정리
