+++ Assets/PMS/MapSystem/PlayerMarker.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public class PlayerMarker : MonoBehaviour
 {
-    //[SerializeField] private Transform player;
+    [SerializeField] private Transform player;
 
     private void Update()
     {
-        //transform.position = new Vector3(player.position.x, transform.position.y, player.position.z);
+        transform.position = new Vector3(player.position.x, transform.position.y, player.position.z);
     }
 
     private void OnCollisionEnter(Collision collision)
