+++ Assets/PMS/Prefabs/PMS_Final_Player.prefab
@@ -1786,7 +1786,7 @@ SkinnedMeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: 5bb68a94905ffc0418608fc7e572ea42, type: 2}
+  - {fileID: 2100000, guid: 360c7fa51bf36ea4ea06f8a9d6adc7f4, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -3030,6 +3030,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 45af6870374407748a935fb3d8a25227, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _fixedWeapon: 0
   _itemType: 6
   _weaponSpawnPos: {fileID: 0}
   _throwData: {fileID: 0}
