+++ Assets/PMS/Scripts/PlayerAnimationScripts/CurrentWeaponCheck.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public class CurrentWeaponCheck : StateMachineBehaviour
 {
     [SerializeField] private WeaponBase _currentWeapon;
-    // OnStateEnter is called when a transition starts and the state machine starts to evaluate this state
+
+    //무기 변경 이벤트 구독시 해제 처리가 애매하다.
+    //그냥 매번 새로운 값을 받아오자.
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        if (_currentWeapon == null)
-        {
-            _currentWeapon = PlayerWeaponManager.Instance.RightCurrentWeapon;
-        }
+        _currentWeapon = PlayerWeaponManager.Instance.RightCurrentWeapon;
+
         if (_currentWeapon != null)
         {
             if (_currentWeapon.ItemType == ItemType.Gun)
