+++ Assets/PMS/Scripts/PlayerAnimationScripts/RifleState/RifleIdleState.cs
@@ -2,41 +2,34 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class RifleAimState : StateMachineBehaviour
+public class RifleIdleState : StateMachineBehaviour
 {
     [SerializeField] private PlayerAttack _playerAttack;
-    private bool flag = true;
+
+    //총을 들고 있으므로 Gun Idle상태가 되어야한다.
+    //만약 playerAttack이 null이면 GetComponent로 들고오자.
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        flag = true;
+        animator.SetLayerWeight(layerIndex, 1f);
         if (_playerAttack == null)
         {
             _playerAttack = animator.gameObject.GetComponent<PlayerAttack>();
         }
     }
 
+    //Idle상태가 유지될 때 
+    //마우스 우클릭을 통해 조준상태 돌입
     public override void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        //이게 자꾸 문제였네
-        if (flag != false) _playerAttack.UpdateAim();
-
-        if (Input.GetMouseButtonDown(0))
-            _playerAttack.RightCurrentWeapon.Attack();
-
-        if (Input.GetMouseButton(1) && flag)
+        if (Input.GetMouseButtonDown(1) && _playerAttack.RightCurrentWeapon.ItemType == ItemType.Gun)
         {
             _playerAttack.ToggleAimMode();
-            animator.SetBool("IsAim", false);
-            flag = false;
         }
-        
     }
 
+    //조준상태 일때 부터는 IsAttacking이 true라서 다른 행동 제약을 줄 수 있음(ex)장비 장착 해제)
     public override void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        if (_playerAttack._rifle != null)
-        {
-            _playerAttack._rifle.isAiming = false;
-        }
+        _playerAttack.IsAttacking = true;
     }
 }
