+++ Assets/PMS/Scripts/Weapon/Gun/Bullet/BulletBase.cs
@@ -5,8 +5,7 @@ using UnityEngine;
 public class BulletBase : MonoBehaviour
 {
     [SerializeField] private int _currentBulletDamage;
-
-    [SerializeField] public float _speed = 10f; //총알 스피드
+    [SerializeField][Range(0.1f,10f)] public float _speed = 10f; //총알 스피드
     [SerializeField] private float _distance = 50f; //총알 유효 사거리
 
     private Rigidbody _rb;
@@ -20,14 +19,14 @@ public class BulletBase : MonoBehaviour
 
     private void OnEnable() //Bullet 객체가 활성화 될 때
     {
-        _startPosition = transform.position; // 활성화될 때마다 시작 위치 저장
-        _rb.velocity = _fireDir * _speed; // 방향은 transform.forward 기준
+        // 활성화될 때마다 시작 위치 저장 -> 거리를 알아서 일정거리 밖으로 나가면 비활성화 처리를 위하여
+        // 방향은 transform.forward 기준
+        _startPosition = transform.position;   
+        _rb.velocity = _fireDir * _speed;       
     }
 
     private void Update()
     {
-        //Move(_fireDir);
-
         //시작점으로 부터 일정 거리 이상멀어지면 비활성화
         if (Vector3.Distance(_startPosition, transform.position) > _distance)
         {
@@ -35,43 +34,38 @@ public class BulletBase : MonoBehaviour
         }
     }
 
-    /*public void Move(Vector3 dir)   
-    { 
-        // 월드 좌표계에서 이동 (로컬 좌표계 문제 해결)
-        transform.position += dir.normalized * _speed * Time.deltaTime;
-    }*/
-
-    // 외부에서 방향을 설정하는 메서드
+    // 외부에서 방향을 설정하는 메서드 -> 의미가 없는거 아닌가?
     public void SetDirection(Vector3 direction)
     {
         _fireDir = direction.normalized;
     }
 
+    // 외부에서 데미지를 받아온다
     public void SetDamage(int damage)
     {
         _currentBulletDamage = damage;
     }
 
+    // IDamageable 컴포넌트 확인
     private void OnCollisionEnter(Collision collision)
-    { 
-        //몬스터 뿐만 아니라 모든 IDamageable을 가지고 있는 모든 객체들
-        if(collision.gameObject.GetComponent<IDamageable>() != null)
+    {
+        //플레이어는 무시
+        if (collision.gameObject.CompareTag("Player")) return;
+
+        //IDamageable을 가지고 있는 객체를 담는다.
+        IDamageable damageable = collision.gameObject.GetComponent<IDamageable>();
+
+        //피해를 받을 객체가 있다면
+        if (damageable != null)
         {
+            //피해 대상자 확인용 Log 디버깅
             Debug.Log(collision.gameObject.name);
-            gameObject.SetActive(false);
-            collision.transform.GetComponent<IDamageable>().Damaged(_currentBulletDamage);
+            damageable.Damaged(_currentBulletDamage);  // 먼저 데미지 처리
+            gameObject.SetActive(false);               // 그 다음 비활성화
         }
-        gameObject.SetActive(false);
-        //벽에 닿으면 없애기
-        /*if (collision.gameObject.CompareTag("Wall"))
+        else
         {
-            gameObject.SetActive(false);
-        }*/
-        //몬스터 닿으면 데미지 처리
-        /*if(collision.gameObject.CompareTag("Monster"))
-        {
-            collision.transform.GetComponent<IDamageable>().Damaged(_currentGunDamage);
-            gameObject.SetActive(false);
-        }*/
+            gameObject.SetActive(false);  // 벽이나 다른 오브젝트와 충돌시 
+        }
     }
 }
