+++ Assets/PMS/Scripts/WeaponBase.cs
@@ -5,25 +5,30 @@ using UnityEngine;
 //최상위 부모 - 공통된 속성은 여기서 처리
 public abstract class WeaponBase : MonoBehaviour
 {
+    //아이템 클래스를 가져온다 -> 나중에 인벤토리에 있는 데이터를 읽고 쓰기 위해서
+    public Item _item;
+    private ItemType _itemType;
+    //플레이어 손위치에 소환도되야하는 위치 local위치를 변경해야 할 것 같음 -> 나중에 고민해보자
+    [SerializeField] protected GameObject _weaponSpawnPos;   
+
+    public ItemType ItemType { get { return _itemType; } protected set { _itemType = value; } }
+    //튜토리얼에 Looting으로 나온 무기들은 Awake.Start,Update 주의 해야한다. 
+    [SerializeField] protected bool _fixedWeapon = false;
+
     // ========== 이벤트 시스템 관련 ==========
     protected bool _canAttack = true; // 공격 가능 상태 저장
-    [SerializeField] protected bool _fixedWeapon = false;
 
     // 컴포넌트가 활성화될 때 이벤트 구독
     protected virtual void OnEnable()
     {
         // PlayerAttack의 이벤트들에 구독
         PlayerAttack.OnAttackStateChanged += HandleAttackStateChanged;
-        PlayerAttack.OnAttackStart += HandleAttackStart;
-        PlayerAttack.OnAttackEnd += HandleAttackEnd;
     }
 
     // 컴포넌트가 비활성화될 때 이벤트 구독 해제 (메모리 누수 방지)
     protected virtual void OnDisable()
     {
         PlayerAttack.OnAttackStateChanged -= HandleAttackStateChanged;
-        PlayerAttack.OnAttackStart -= HandleAttackStart;
-        PlayerAttack.OnAttackEnd -= HandleAttackEnd;
     }
 
     // 공격 상태 변경 이벤트 핸들러
@@ -33,12 +38,6 @@ public abstract class WeaponBase : MonoBehaviour
         OnAttackStateChanged(canAttack);
     }
 
-    // 공격 시작 이벤트 핸들러
-    private void HandleAttackStart()
-    {
-        OnAttackStarted();
-    }
-
     //공격 상태 변경 시점
     // 하위 클래스에서 오버라이드할 수 있는 가상 메서드들
     protected virtual void OnAttackStateChanged(bool canAttack)
@@ -46,35 +45,6 @@ public abstract class WeaponBase : MonoBehaviour
         // 필요한 경우 하위 클래스에서 구현
         //Debug.Log($"{gameObject.name}: 공격 가능 상태 변경 - {canAttack}");
     }
-
-    //공격 시작 했을 때
-    protected virtual void OnAttackStarted()
-    {
-        // 필요한 경우 하위 클래스에서 구현
-        //Debug.Log($"{gameObject.name}: 공격 시작됨");
-    }
-
-    //공격 끝났을 때
-    protected virtual void OnAttackEnded()
-    {
-        // 필요한 경우 하위 클래스에서 구현
-        //Debug.Log($"{gameObject.name}: 공격 종료됨");
-    }
-
-    // 공격 종료 이벤트 핸들러
-    private void HandleAttackEnd()
-    {
-        OnAttackEnded();
-    }
-
-    //아이템 클래스를 가져온다 -> 나중에 인벤토리에 있는 데이터를 읽고 쓰기 위해서
-    public Item _item;      
-
-    [SerializeField] protected ItemType _itemType;
-    [SerializeField] protected GameObject _weaponSpawnPos;    //플레이어 손위치에 소환도되야하는 위치 local위치를 변경해야 할 것 같음
-
-    public ItemType ItemType { get { return _itemType; } protected set {_itemType = value; } }
-    public virtual void Defense() { }   //쉴드방어
 
     //각 무기에서 초기화 시 더필요한 부분 override하고 base.Init()
     public virtual void Init()
@@ -93,7 +63,7 @@ public abstract class WeaponBase : MonoBehaviour
                 gameObject.transform.localRotation = _weaponSpawnPos.transform.localRotation;
             }
         }
-    } //무기 초기화 함수
+    } 
 
     // Attack 메서드 수정 - 공격 가능 상태 체크 추가
     public virtual void Attack()
